# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "Guardian"
  description: "The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation."
  version: "2.9.0"
  contact:
    name: "API developer"
    url: "https://envisionblockchain.com"
    email: "info@envisionblockchain.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
- url: /api/v1
  description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
    Credentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    Account:
      type: object
      required:
        - username
        - role
      properties:
        username:
          type: string
        role:
          type: string
        did:
          type: string
    Session:
      type: object
      required:
        - username
        - role
        - accessToken
      properties:
        username:
          type: string
        role:
          type: string
        accessToken:
          type: string
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
paths:
  /accounts:
    get:
      tags:
        - accounts
      description: Returns all users except those with roles Standard Registry and Auditor. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Returns a list of users, excluding Standard Registry and Auditors.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/session:
    get:
      tags:
        - accounts
      description: Returns current user session.
      security:
      - bearerAuth: []
      summary: Returns current session of the user.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/balance:
    get:
      tags:
        - accounts
      description: Requests current Hedera account balance.
      security:
      - bearerAuth: []
      summary: Returns user's Hedera account balance.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                  unit:
                    type: string
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                      did:
                        type: string
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/register:
    post:
      tags:
        - accounts
      summary: Registers a new user account.
      requestBody:
        description: Object that contain username, password and role (optional) fields.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        201:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/login:
    post:
      tags:
        - accounts
      summary: Logs user into the system.
      requestBody:
        description: Object that contains username and password fields.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/root-authorities:
    get:
      deprecated: true
      tags:
        - accounts
      description: Returns all Standard Registries.
      security:
      - bearerAuth: []
      summary: Returns all Standard Registries.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/standard-registries:
    get:
      tags:
        - accounts
      description: Returns all Standard Registries.
      security:
      - bearerAuth: []
      summary: Returns all Standard Registries.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...